Add files via upload
The RFID Based Vehicle Starter System is a security project that allows a vehicle to start only when an authorized RFID card or tag is scanned. The system uses an RC522 RFID reader to identify the card, an Arduino UNO to process the UID, and a relay to control the engine ignition. A 16×2 LCD displays messages like Access Granted or Access Denied. This ensures safe, touchless, and password-free vehicle access.